+++ CHANGELOG.md
@@ -80,3 +80,4 @@ All notable changes to `strapi-wrapper` will be documented in this file.
 ### 0.2.8
 
 - initial StrapiImage handler
+- Collection Update and delete